# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ampara Animal — contrato do Ciclo 1",
    "version": "0.3.0",
    "description": "Contrato do Ciclo 1 (Sprint 3) e do prontuário do Ciclo 2. Endpoints marcados x-situacao=implementado existem no código e são conferidos por scripts/validar-openapi.cjs. Todas as rotas exigem login (Authorization: Bearer <access token do Supabase Auth>), exceto /api/health e /api/health/ready. IDs e dados dos exemplos são sintéticos. Autorização: além do login, a conta precisa estar liberada na API (tabela Usuario); conta não liberada ou desativada recebe 403. ADMIN acessa todas as unidades; OPERADOR só as vinculadas (papéis provisórios até aprovação da Ampara). O escopo é decidido no servidor: filtros enviados pelo cliente nunca ampliam o acesso."
  },
  "paths": {
...
    "/api/v1/fotos": {
      "post": {
        "summary": "Solicitar envio de foto",
        "description": "Implementado. Gera uma URL assinada temporária (15 min) para envio direto do arquivo ao armazenamento. 409 se o id já pertence a uma foto confirmada (gere um novo id para outra foto).",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "201": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvioFoto"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou foto enviada por outro usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FotoSolicitada"
              }
            }
          }
        }
      }
    },
    "/api/v1/fotos/{id}/confirmacao": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identificador do recurso",
          "schema": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "summary": "Confirmar existência e validade do arquivo enviado",
        "description": "Implementado. Verifica se o arquivo chegou ao armazenamento e marca a foto como confirmada. Confere o arquivo real no Storage: acima de 8 MB ou tipo diferente do declarado retorna 400 e o arquivo é removido.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FotoConfirmada"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou foto enviada por outro usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Erro": {
        "type": "object",
        "required": [
          "statusCode",
          "mensagem",
          "caminho",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "mensagem": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "caminho": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          },
          "detalhes": {
            "type": "object",
            "additionalProperties": true,
            "description": "Opcional, só em alguns 4xx. Ex.: 409 de microchip duplicado traz animalExistenteId e identificadorPublico para o frontend oferecer a transferência."
          }
        },
        "example": {
          "statusCode": 409,
          "mensagem": "Versão desatualizada",
          "caminho": "/api/v1/animals/56b9f220-3c58-4adc-9dfa-102b90720e12",
          "timestamp": "2026-09-13T12:00:00Z"
        }
      },
...
      "FotoSolicitada": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tipoMidia",
          "tamanhoBytes"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "tipoMidia": {
            "type": "string",
            "description": "Tipos aceitos ainda sujeitos a acordo",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "tamanhoBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8388608
          }
        },
        "example": {
          "id": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "tipoMidia": "image/jpeg",
          "tamanhoBytes": 123456
        }
      },
      "EnvioFoto": {
        "type": "object",
        "required": [
          "id",
          "urlEnvio",
          "expiraEm"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "urlEnvio": {
            "type": "string",
            "description": "URL assinada temporária para PUT binário; não registrar em logs",
            "format": "uri"
          },
          "expiraEm": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          }
        },
        "example": {
          "id": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "urlEnvio": "https://example.invalid/envio",
          "expiraEm": "2026-09-13T12:15:00Z"
        }
      },
      "FotoConfirmada": {
        "type": "object",
        "required": [
          "id",
          "situacao"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "situacao": {
            "type": "string",
            "enum": [
              "CONFIRMADA"
            ]
          }
        },
        "example": {
          "id": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "situacao": "CONFIRMADA"
        }
      },
...
    },
    "securitySchemes": {
      "bearerSupabase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token emitido pelo Supabase Auth do projeto (o mesmo login do frontend). Validado no servidor pela assinatura (JWKS), emissor e audiência \"authenticated\". Sem token ou com token inválido/vencido: 401."
      }
    }
  }
}